+++ src/main/resources/templates/presentationGenerator/UiState.kt
@@ -1,9 +1,0 @@
-package ${PACKAGE}
-
-import androidx.compose.runtime.Immutable
-
-@Immutable
-internal data class ${SCREEN_NAME}UiState(
-    val isLoading: Boolean = false,
-    val error: String? = null,
-)